Update uptime field type in HealthCheckResponse struct from u128 to u64

+++ src/server/dto.rs
@@ -5,7 +5,7 @@ use serde::{Deserialize, Serialize};
 pub struct HealthCheckResponse {
     pub status: String,
     pub service: String,
-    pub uptime: u128,
+    pub uptime: u64,
 }
 
 /// Request DTO for database creation
